Move backend auth logic to seperate file
